+++ HISTORY.rst
@@ -16,3 +16,14 @@ History
 ------------------
 
 * Travis bug fix
+
+0.2.0 (2020-07-13)
+------------------
+
+* Add columns to guide and gene output for base LFC of pairs
+* Check inputs, removing guides without the right number of pairs or control pairs
+
+
+0.2.1 (2020-07-13)
+------------------
+* Update basic usage notebook